telegram_bot/requirements.txt
python-telegram-bot>=21.0
openai>=1.50.0
httpx>=0.27.0
python-dotenv>=1.0.0
